fix logic at message processing
